requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-database-view==0.3.0
django-filter==21.1
djangorestframework==3.12.4
django-rest-swagger==2.1.2
docopt==0.6.2
filelock==3.3.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
heroku==0.1.4
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Markdown==3.3.4
pipenv==2021.11.9
pipreqs==0.4.11
platformdirs==2.4.0
psycopg2==2.9.1
python-dateutil==1.5
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
yarg==0.1.9
drf-yasg==1.20.0
djangorestframework-simplejwt==5.1.0
pyjwt==2.3.0
Pillow==9.1.0
moviepy==1.0.3